chore!: renamed `checks` to `courses` in `hooks`, `lib`
Renamed the `checks` directory to `courses` for both `hooks` and `lib`.
