+++ database/course/update.ts
@@ -8,7 +8,7 @@ import { updateSettings } from '@/database/course/settings/update'
 import getDatabase, { DrizzleDB } from '@/database/Database'
 import { db_course } from '@/database/drizzle/schema'
 import requireAuthentication from '@/src/lib/auth/requireAuthentication'
-import { LodashDifferences } from '@/src/lib/checks/create/SaveAction'
+import { LodashDifferences } from '@/src/lib/courses/create/SaveAction'
 import _logger from '@/src/lib/log/Logger'
 import { Course, CourseSchema } from '@/src/schemas/KnowledgeCheck'
 import createConvertToDatabase from '@/src/schemas/utils/createConvertToDatabase'
