+++ src/app/[locale]/results/[courseId]/practice/page.tsx
