+++ src/components/courses/[share_token]/(shared)/(questions)/OpenQuestionAnswer.tsx
@@ -3,7 +3,7 @@ import { MessageCircleQuestionIcon } from 'lucide-react'
 import { useFormContext } from 'react-hook-form'
 import DisplayFeedbackText from '@/src/components/courses/[share_token]/practice/FeedbackText'
 import { Textarea } from '@/src/components/shadcn/textarea'
-import { usePracticeFeeback } from '@/src/hooks/checks/[share_token]/practice/usePracticeFeedback'
+import { usePracticeFeeback } from '@/src/hooks/courses/[share_token]/practice/usePracticeFeedback'
 import { useRHFContext } from '@/src/hooks/Shared/form/react-hook-form/RHFProvider'
 import { cn } from '@/src/lib/Shared/utils'
 import type { OpenQuestion } from '@/src/schemas/QuestionSchema'
