+++ src/components/courses/[share_token]/RenderExamQuestion.tsx
@@ -9,7 +9,7 @@ import { Button } from '@/src/components/shadcn/button'
 import { Form } from '@/src/components/shadcn/form'
 import DragDropContainer from '@/src/components/Shared/drag-drop/DragDropContainer'
 import { DragDropItem } from '@/src/components/Shared/drag-drop/DragDropItem'
-import { ExaminationActions } from '@/src/hooks/checks/[share_token]/ExaminationStore'
+import { ExaminationActions } from '@/src/hooks/courses/[share_token]/ExaminationStore'
 import debounceFunction from '@/src/hooks/Shared/debounceFunction'
 import { ChoiceQuestion, DragDropQuestion } from '@/src/schemas/QuestionSchema'
 import { QuestionInput, QuestionInputSchema } from '@/src/schemas/UserQuestionInputSchema'
