+++ src/components/courses/[share_token]/practice/RenderPracticeQuestion.tsx
@@ -14,11 +14,11 @@ import DisplayFeedbackText from '@/src/components/courses/[share_token]/practice
 import { usePracticeStore } from '@/src/components/courses/[share_token]/practice/PracticeStoreProvider'
 import { Button } from '@/src/components/shadcn/button'
 import FormFieldError from '@/src/components/Shared/form/FormFieldError'
-import { usePracticeFeeback } from '@/src/hooks/checks/[share_token]/practice/usePracticeFeedback'
+import { usePracticeFeeback } from '@/src/hooks/courses/[share_token]/practice/usePracticeFeedback'
 import { useLogger } from '@/src/hooks/log/useLogger'
 import { RHFProvider, useRHFContext } from '@/src/hooks/Shared/form/react-hook-form/RHFProvider'
 import useRHF from '@/src/hooks/Shared/form/useRHF'
-import { EvaluateAnswer } from '@/src/lib/checks/[share_token]/practice/EvaluateAnswer'
+import { EvaluateAnswer } from '@/src/lib/courses/[share_token]/practice/EvaluateAnswer'
 import { cn } from '@/src/lib/Shared/utils'
 import { ChoiceQuestion, Question } from '@/src/schemas/QuestionSchema'
 import { QuestionInput, QuestionInputSchema } from '@/src/schemas/UserQuestionInputSchema'
