+++ src/lib/courses/[share_token]/isExaminationAllowed.ts
