+++ src/lib/courses/computeQuestionScore.ts
