+++ src/lib/courses/hasCollaborativePermissions.ts
