+++ src/tests/integration/DatabaseInsertionOrder.test.ts
@@ -4,7 +4,7 @@ import insertCourse from '@/database/course/insert'
 import { getCourseById } from '@/database/course/select'
 import getDatabase from '@/database/Database'
 import { db_course, db_user } from '@/database/drizzle/schema'
-import prepareExaminationCheck from '@/src/lib/checks/[share_token]/prepareExminationCheck'
+import prepareExaminationCheck from '@/src/lib/courses/[share_token]/prepareExminationCheck'
 import { Course, instantiateCourse } from '@/src/schemas/KnowledgeCheck'
 import { CourseSettings } from '@/src/schemas/KnowledgeCheckSettingsSchema'
 import { instantiateDragDropQuestion, instantiateMultipleChoice, instantiateOpenQuestion, instantiateSingleChoice } from '@/src/schemas/QuestionSchema'
